weight in int type

+++ bmiCalculator.py
@@ -1,5 +1,5 @@
 height = float(input("please insert your height in meter units: \n"))
-weight = float(input("please insert your weight in kg units: \n"))
+weight = int(input("please insert your weight in kg units: \n"))
 bmi = weight / (height ** 2)
 
 print("yout BMI is: ", bmi)
